requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
numpy>=1.24.3
plotly>=5.14.1
snowflake-connector-python>=3.0.0
python-dotenv>=1.0.0
